project refactior

+++ src/main/java/com/objetivo/repository/EnderecoRepository.java
@@ -1,5 +1,6 @@
 package com.objetivo.repository;
 
+import com.objetivo.entities.Pessoa;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
@@ -7,11 +8,14 @@ import org.springframework.stereotype.Repository;
 
 import com.objetivo.entities.Endereco;
 
+import java.util.List;
+import java.util.Locale;
+
 @Repository
-public interface EnderecoRepository  extends JpaRepository<Endereco, Integer>{
+public interface EnderecoRepository  extends JpaRepository<Endereco, Long>{
 	
 	@Query(nativeQuery = true, value = "select e.pessoa_id from elo.endereco e where e.id = :id")
-	Integer findPessoaByEndereco(@Param("id") Integer id);
+	Long findPessoaByEndereco(@Param("id") Long id);
 	
 	@Query(nativeQuery = true, value = "select max(p.id) from elo.pessoa p")
 	Integer findMaxPessoa();
